api/requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
openai>=1.55.0
pydantic>=2.10.0
supabase>=2.10.0
upstash-redis>=1.0.0
python-dotenv>=1.0.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.111.0
requests>=2.31.0
pypdf>=3.17.0
python-docx>=1.1.0
